remap buttons

+++ radio.py
@@ -8,6 +8,7 @@ import time
 import signal
 from io import BytesIO
 import threading
+import random
 
 streams = {
     'KQED': {
@@ -48,12 +49,6 @@ streams = {
     }
 }
 
-button_mappings = {
-    'A':'NTS 2',
-    'B':'NTS 1',
-    'X':'KQED',
-    'Y':'Hyde FM'
-}
 
 disp = st7789.ST7789(
     height=240,
@@ -67,23 +62,8 @@ disp = st7789.ST7789(
 )
 
 disp.begin()
-#image = Image.new('RGB', (240, 240), color=(0, 0, 0))
-#draw = ImageDraw.Draw(image)
-
-#for button, name in button_mappings.items():
-#    logo_path = streams[name]['logo']
-#    logo_path = f'logos/{logo_path}'
-#    border = Image.new('RGB', (57, 57), color=(255, 255, 255))
-#    logo = Image.open(logo_path).resize((55, 55))
-#
-#    if button=='A':
-#        image.paste(border, (3, 3))
-#        image.paste(logo, (5, 5))
-#    elif button=='B':
-#        image.paste(border, (240-57-3, 3))
-#        image.paste(logo, (240-55-5, 5))
-
-#disp.display(image.rotate(180))
+image = Image.new('RGB', (240, 240), color=(0, 0, 0))
+disp.display(image)
 
 mpv_process = None
 stream = None
@@ -195,17 +175,30 @@ def display_info(name, play_status):
 def toggle_stream(name):
     global mpv_process, stream
 
-    stream_info = streams[name]
-    stream_url = stream_info['stream']
+    if name != None:
 
-    if mpv_process: # if stream is playing, stop it
-        mpv_process.send_signal(signal.SIGTERM)
-        mpv_process = None
+        stream_info = streams[name]
+        stream_url = stream_info['stream']
 
-        display_info(name, 'pause')
+        if mpv_process: # if stream is playing, stop it
+            mpv_process.send_signal(signal.SIGTERM)
+            mpv_process = None
 
-        if stream != name: # if the button pressed is a new stream, play it
-            mpv_process = Popen([ 
+            display_info(name, 'pause')
+
+            if stream != name: # if the button pressed is a new stream, play it
+                mpv_process = Popen([ 
+                    "mpv",
+                    "--ao=alsa",
+                    "--audio-device=alsa/hw:1,0",
+                    "--volume=50",
+                    stream_url
+                ])
+                stream = name
+                display_info(name, 'play')
+
+        else: # otherwise play the one pressed
+            mpv_process = Popen([
                 "mpv",
                 "--ao=alsa",
                 "--audio-device=alsa/hw:1,0",
@@ -215,17 +208,27 @@ def toggle_stream(name):
             stream = name
             display_info(name, 'play')
 
-    else: # otherwise play the one pressed
-        mpv_process = Popen([
-            "mpv",
-            "--ao=alsa",
-            "--audio-device=alsa/hw:1,0",
-            "--volume=50",
-            stream_url
-        ])
-        stream = name
-        display_info(name, 'play')
-        
+def play_random():
+    available_streams = [i for i in list(streams.keys()) if i != stream]
+    chosen = random.choice(available_streams)
+    toggle_stream(chosen)
+
+
+def seek_stream(direction):
+    if stream == None:
+        play_random()
+    
+    else:
+        stream_list = list(streams.keys())
+        idx = stream_list.index(stream)
+        try:
+            toggle_stream(stream_list[idx + direction])
+        except:
+            if direction == 1:
+                toggle_stream(stream_list[0])
+            else:
+                toggle_stream(stream_list[-1])
+
 def shutdown():
     run(['sudo', 'shutdown', 'now'])
 
@@ -239,10 +242,10 @@ button_y = Button(24, hold_time=5)
 button_a = Button(5, hold_time=5)
 button_b = Button(6, hold_time=5)
 
-button_b.when_pressed = lambda: toggle_stream('NTS 1')
-button_a.when_pressed = lambda: toggle_stream('NTS 2')
-button_y.when_pressed = lambda: toggle_stream('HydeFM')
-button_x.when_pressed = lambda: toggle_stream('Dublab')
+button_b.when_pressed = lambda: play_random()
+button_a.when_pressed = lambda: toggle_stream(stream)
+button_y.when_pressed = lambda: seek_stream(-1)
+button_x.when_pressed = lambda: seek_stream(1)
 
 periodic_update()
